refactor to abstract common nav Controller in an abstract class
