+++ app/src/org/commcare/activities/NavigationHostCommCareActivity.java
@@ -1,0 +1,39 @@
+package org.commcare.activities;
+
+import android.os.Bundle;
+
+import androidx.navigation.NavController;
+import androidx.navigation.fragment.NavHostFragment;
+
+import org.commcare.dalvik.R;
+import org.commcare.google.services.analytics.FirebaseAnalyticsUtil;
+
+import javax.annotation.Nullable;
+
+public abstract class NavigationHostCommCareActivity<R> extends CommCareActivity<R>{
+
+    private NavController.OnDestinationChangedListener destinationListener;
+
+    protected NavController navController;
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(getLayoutResource());
+        destinationListener = FirebaseAnalyticsUtil.getNavControllerPageChangeLoggingListener();
+        navController = getHostFragment().getNavController();
+        navController.addOnDestinationChangedListener(destinationListener);
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        NavController navController = getHostFragment().getNavController();
+        navController.removeOnDestinationChangedListener(destinationListener);
+        destinationListener = null;
+    }
+
+   protected abstract int getLayoutResource();
+
+    protected abstract NavHostFragment getHostFragment();
+}
